+++ controller/color/index.js
@@ -1,10 +1,20 @@
 const colors = require("../../models").colors;
 const { colorValidator } = require("../../validator/color");
+const { Op } = require("sequelize");
 
 exports.getColors = async function (req, res) {
+  const { limit, page, name } = req.body;
+
   try {
-    const color = await colors.findAll();
-    res.send(color);
+    if (!name) {
+      res.status(400).send("Please pass name in your body request");
+    } else {
+      const color = await colors.findAndCountAll({
+        limit: limit,
+        where: { name: { [Op.like]: `%${name}%` } },
+      });
+      res.send(color);
+    }
   } catch (error) {
     res.status(400).send(`Something went Wrong: ${error}`);
   }
@@ -16,15 +26,6 @@ exports.createColor = async function (req, res) {
     if (error) return res.status(400).send(error.details[0].message);
 
     const { name, hex_code } = req.body;
-
-    if (
-      !{
-        name,
-        hex_code,
-      }
-    ) {
-      res.status(400).send("All input is required");
-    }
 
     const color = await colors.create({
       name,
@@ -40,7 +41,7 @@ exports.getColorById = async function (req, res) {
   const id = req.params.id;
   try {
     if (!id) res.status(400).send("ID is required!");
-    const color = await colors.findOne({ where: { id: String(id) } });
+    const color = await colors.findOne({ where: { id: id } });
     res.status(200).json(color);
   } catch (error) {
     res.status(400).send(`Something went Wrong: ${error}`);
@@ -50,12 +51,21 @@ exports.getColorById = async function (req, res) {
 exports.deleteColor = async function (req, res) {
   const id = req.params.id;
   try {
-    const color = await colors.destroy({
-      where: {
-        id: String(id),
-      },
-    });
-    res.status(201).send(`color with ${id} deleted successfully!`);
+    const oldColor = await colors.findOne({ where: { id: id } });
+
+    if (oldColor) {
+      const color = await colors.destroy({
+        where: {
+          id: id,
+        },
+      });
+      res.status(200).send(`color with ${id} deleted successfully!`);
+      console.log("ID", oldColor);
+    } else {
+      res
+        .status(400)
+        .send(`color with ${id} didnt find! please send available ID`);
+    }
   } catch (error) {
     res.status(400).send(`Something went Wrong: ${error}`);
   }
@@ -63,7 +73,7 @@ exports.deleteColor = async function (req, res) {
 
 exports.updateColor = async function (req, res) {
   const id = req.params.id;
-  const { name, family, phone, email, address } = req.body;
+  const { name, hex_code } = req.body;
 
   try {
     const color = await colors.update(
