+++ controller/item/index.js
@@ -1,10 +1,20 @@
 const items = require("../../models").items;
 const { itemValidator } = require("../../validator/item");
+const { Op } = require("sequelize");
 
 exports.getItems = async function (req, res) {
+  const { limit, page, name } = req.body;
+
   try {
-    const item = await items.findAll();
-    res.send(item);
+    if (!name) {
+      res.status(400).send("Please pass name in your body request");
+    } else {
+      const item = await items.findAndCountAll({
+        limit: limit,
+        where: { name: { [Op.like]: `%${name}%` } },
+      });
+      res.send(item);
+    }
   } catch (error) {
     res.status(400).send(`Something went Wrong: ${error}`);
   }
@@ -17,10 +27,6 @@ exports.createItem = async function (req, res) {
 
     const { name, image } = req.body;
 
-    if (!{ name, image }) {
-      res.status(400).send("All input is required");
-    }
-
     const item = await items.create({ name, image });
     res.status(201).json(item);
   } catch (error) {
@@ -32,7 +38,7 @@ exports.getItemById = async function (req, res) {
   const id = req.params.id;
   try {
     if (!id) res.status(400).send("ID is required!");
-    const item = await items.findOne({ where: { id: String(id) } });
+    const item = await items.findOne({ where: { id: id } });
     res.status(200).json(item);
   } catch (error) {
     res.status(400).send(`Something went Wrong: ${error}`);
@@ -42,12 +48,21 @@ exports.getItemById = async function (req, res) {
 exports.deleteItem = async function (req, res) {
   const id = req.params.id;
   try {
-    const item = await items.destroy({
-      where: {
-        id: String(id),
-      },
-    });
-    res.status(201).send(`item with ${id} deleted successfully!`);
+    const oldItem = await items.findOne({ where: { id: id } });
+
+    if (oldItem) {
+      const item = await items.destroy({
+        where: {
+          id: String(id),
+        },
+      });
+      res.status(200).send(`item with ${id} deleted successfully!`);
+      console.log("ID", oldItem);
+    } else {
+      res
+        .status(400)
+        .send(`item with ${id} didnt find! please send available ID`);
+    }
   } catch (error) {
     res.status(400).send(`Something went Wrong: ${error}`);
   }
@@ -60,10 +75,10 @@ exports.updateItem = async function (req, res) {
   try {
     const item = await items.update(
       {
-        name: name,
-        brand_id: brand_id,
-        color_id: color_id,
-        subtitle: subtitle,
+        name,
+        brand_id,
+        color_id,
+        subtitle,
       },
       { returning: true, where: { id: id } }
     );
